CancerRAG/frontend/requirements.txt
langserve==0.3.0
httpx==0.27.2
streamlit
python-dotenv
pydantic
pyaml
fastapi==0.115.4
textstat==0.7.4
vaderSentiment
nltk